Switching the logic to better handle non-pending processes.

+++ grow/preprocessors/gulp.py
@@ -56,23 +56,22 @@ class GulpPreprocessor(base.BasePreprocessor):
 @atexit.register
 def _kill_child_process():
     """Sometimes the child process keeps going after grow is done running."""
+    has_pending = False
     for process in _child_processes:
         try:
-            process.terminate()
+            if process.poll() is None:
+                has_pending = True
+                process.terminate()
         except OSError:
             # Ignore the error.  The OSError doesn't seem to be documented(?)
             pass
 
+    if not has_pending:
+        return
+
     # Give the process some time to finish by itself.
+    time.sleep(2)
 
-    while True:
-        time.sleep(1)
-
-        is_pending = False
-        for process in _child_processes:
-            if process.poll() is None:
-                is_pending = True
-                process.kill()
-
-        if not is_pending:
-            return
+    for process in _child_processes:
+        if process.poll() is None:
+            process.kill()
